requirements.txt
python-escpos==3.0a9
openai>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
requests>=2.28.0
python-dateutil>=2.8.0
colorama>=0.4.0
pydantic>=2.0.0
flask>=2.0.0
flask-cors>=4.0.0
flask-socketio>=5.0.0
python-socketio>=5.0.0
werkzeug>=2.0.0
bcrypt>=4.0.0